I'm an idiot and cast float to float instead of float to double

+++ typeClassDefs/WALK2014_ControlStatus.h
@@ -53,10 +53,10 @@ namespace guWhiteboard
                 /** comparison operator */
 		inline bool operator == (const WALK2014_ControlStatus &s) 
 		{
-    			return (fabs(static_cast<float>(forward() - s.forward())) < DBL_EPSILON &&
-				fabs(static_cast<float>(left() - s.left())) < DBL_EPSILON &&
-				fabs(static_cast<float>(turn() - s.turn())) < DBL_EPSILON &&
-				fabs(static_cast<float>(power() - s.power())) < DBL_EPSILON &&
+    			return (fabs(static_cast<double>(forward() - s.forward())) < DBL_EPSILON &&
+				fabs(static_cast<double>(left() - s.left())) < DBL_EPSILON &&
+				fabs(static_cast<double>(turn() - s.turn())) < DBL_EPSILON &&
+				fabs(static_cast<double>(power() - s.power())) < DBL_EPSILON &&
 				foot() == s.foot() &&
 				walk_state()	== s.walk_state() &&
 				walking_status()	== s.walking_status() &&
